# improved speed of fast corner detect