@lagru 2024-04-15
Thank you! It may be some time until someone of us get's to review this, performance optimization aren't the highest priority right now. But it will get done eventually. :pray: