@spdfghi 2024-04-17
> May I ask for a test or benchmark as a companion to this change? 🙏
@mkcor in #7388 I made a simple test on a constant image and achieved results with over a fifty percent reduction in time. Essentially, this is equivalent to redundant measurements at non-corner positions. If you have any other thoughts, please let me know.
here is the code:
```
import numpy as np
import timeit
from skimage.feature import corner_fast, corner_peaks
square = np.zeros((1200, 1200))
square.astype(int)
def test():
corner_fast(square)

execution_time = timeit.timeit(test, number=100)
print(execution_time)
```