@lagru 2024-04-24
Okay, using 

```python
import numpy as np
import skimage as ski
from skimage.feature import corner_fast

zeros = np.zeros((1000, 1000))
hubble = ski.color.rgb2gray(ski.data.hubble_deep_field())

%timeit corner_fast(zeros, 13)
%timeit corner_fast(hubble, 13)
```
this optimization seems faster indeed:

|   | zeros | hubble |
|--------|--------|--------|
| `main` | 17.5 ms ± 13 µs | 17.1 ms ± 158 µs |
| #7394 | 5.84 ms ± 14.5 µs | 6.97 ms ± 11.8 |