@lagru 2024-04-24
Updated the comment above. The optimization is slightly slower, for cases where there are a lot of potential corners. I suspect this is because `ring_pixel = image[i + rp[k], j + cp[k]]` and its comparisons are wasted. Maybe we can figure out an implementation that re-uses `ring_pixel`?