@spdfghi 2024-04-24
> Updated the comment above. The optimization is slightly slower, for cases where there are a lot of potential corners. I suspect this is because `ring_pixel = image[i + rp[k], j + cp[k]]` and its comparisons are wasted. Maybe we can figure out an implementation that re-uses `ring_pixel`?



@lagru I believe it could be done by assigning these four variables first and then manually assigning the other eight, without using a for loop. However, is it necessary to do so? For the random images you are using, the number of feature points should be around 1/8 of the total number of image pixels, as determined by simple mathematical calculations. However, for real images, the number of feature points is generally less than 1/100. Additionally, for commonly used feature matching methods such as k-nearest neighbors (knn), the time complexity is O(N^2), which requires us to control the number of feature points. For an image of size 600*800, if the total number of feature points is around 1/8, the matching time will be on the order of 1e9, which is clearly far beyond the feature extraction time and is unacceptable.