@lagru 2024-04-30
In the [paper](http://www.edwardrosten.com/work/rosten_2006_machine.pdf) it even says

> Knowledge from the first 4 tests is discarded.

So all in all, I agree with @spdfghi. Let's not over-optimize what may very well be an edge case. This should be an improvement for most cases and also align more with the cited source.